+++ src/PluginManager/include/PluginManager/Location/LocationExportPluginBaseSettings.h
@@ -46,8 +46,8 @@ public:
     LocationExportPluginBaseSettings &operator=(LocationExportPluginBaseSettings &&rhs) = delete;
     ~LocationExportPluginBaseSettings() override;
 
-    bool isExportSystemLocationsEnabled() const noexcept;
-    void setExportSystemLocationsEnabled(bool enabled) noexcept;
+    bool isExportPresetLocationsEnabled() const noexcept;
+    void setExportPresetLocationsEnabled(bool enabled) noexcept;
 
     bool isOpenExportedFilesEnabled() const noexcept;
     void setOpenExportedFilesEnabled(bool enabled) noexcept;